Moves to a simpler package name
