+++ demo/src/main/java/com/novoda/downloadmanager/Authority.java
@@ -1,4 +1,4 @@
-package com.novoda.downloadmanager.lib;
+package com.novoda.downloadmanager;
 
 public class Authority {
     public static final String AUTHORITY = "com.novoda.downloadmanager.demo";
